Deprecate pytz and use zoneinfo (#8159)
* Initial test coverage update

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Fully remove pytz

* remove from dependencies

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* bug fix

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* test fixes

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix app test

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* noqa

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix

* small change

* Add tzdata for windows

* Test case

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Fix formatting

* Improved documentation

* Fix

* remove

* Fix

* Fix

* more accurate

* Comment

* docstrings

* future import

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* pre-commit

* Fix

* fix

* docstring fix

* comment

* pre-commit

* trailing whitespace fix

* Update documentation

* Update celery/utils/time.py

* Update celery/utils/time.py

* Update celery/utils/time.py

* Update celery/utils/time.py

* Update celery/utils/time.py

---------

+++ celery/utils/text.py
@@ -93,7 +93,7 @@ def truncate(s: str, maxlen: int = 128, suffix: str = '...') -> str:
     return s
 
 
-def pluralize(n: int, text: str, suffix: str = 's') -> str:
+def pluralize(n: float, text: str, suffix: str = 's') -> str:
     """Pluralize term when n is greater than one."""
     if n != 1:
         return text + suffix
